Add lowering from thunk AST to MLIR AST
MLIR is just a mid-level intermediate representation. It reduces the graph by removing movement ops via merging views. We only have one view right now but this will need to change. MLIR also holds the args for ops (buffs for loads and stores and axis for reduce) and the dtype. print_ast is updated for MLIR since thunk trees are not helpful at this stage.